Fix around numeric type for AVG

+++ community/cypher/src/main/scala/org/neo4j/cypher/internal/pipes/aggregation/AvgFunction.scala
@@ -35,7 +35,7 @@ class AvgFunction(val value: Expression)
 
   def result =
     if (count > 0)
-      divide(sofar, count)
+      divide(sofar, count.toDouble)
     else
       null
